Upgrading maven archetype project for jar applications : - config - changing first version of database scripts (001 instead of 002) - config - adding the welcome message (StringTemplate) - jar - adding settings properties file to the SessionController constructor call - war - completing content of RequestRouter class - war - adding CSS file - war - adding DOCTYPE to the HTML page
