+++ jar-component-archetype/pom.xml
@@ -35,7 +35,6 @@
   
   <groupId>com.silverpeas.archetypes</groupId>
   <artifactId>jar-component-archetype</artifactId>
-  <version>${parent.version}</version>
   <packaging>maven-archetype</packaging>
-  <name>Silverpeas Archetype with Jar component</name>
+  <name>Silverpeas Archetype with Jar application</name>
 </project>
