+++ jar-component-archetype/src/main/resources/archetype-resources/__rootArtifactId__-config/src/main/config/dbRepository/data/mssql/__rootArtifactId__-contribution.xml
@@ -32,24 +32,24 @@
     <requirement>
     </requirement>
 
-    <current version="002">
+    <current version="001">
         <create_table>
-            <file name="mssql\\${parentArtifactId}\002\create_table.sql" type="sqlstatementlist" delimiter=";" keepdelimiter="NO"/>
+            <file name="mssql\\${parentArtifactId}\001\create_table.sql" type="sqlstatementlist" delimiter=";" keepdelimiter="NO"/>
         </create_table>
         <create_constraint>
-            <file name="mssql\\${parentArtifactId}\002\create_constraint.sql" type="sqlstatementlist" delimiter=";" keepdelimiter="NO"/>
+            <file name="mssql\\${parentArtifactId}\001\create_constraint.sql" type="sqlstatementlist" delimiter=";" keepdelimiter="NO"/>
         </create_constraint>
         <create_index>
-            <file name="mssql\\${parentArtifactId}\002\create_index.sql" type="sqlstatementlist" delimiter=";" keepdelimiter="NO"/>
+            <file name="mssql\\${parentArtifactId}\001\create_index.sql" type="sqlstatementlist" delimiter=";" keepdelimiter="NO"/>
         </create_index>
         <drop_table>
-            <file name="mssql\\${parentArtifactId}\002\drop_table.sql" type="sqlstatementlist" delimiter=";" keepdelimiter="NO"/>
+            <file name="mssql\\${parentArtifactId}\001\drop_table.sql" type="sqlstatementlist" delimiter=";" keepdelimiter="NO"/>
         </drop_table>
         <drop_constraint>
-            <file name="mssql\\${parentArtifactId}\002\drop_constraint.sql" type="sqlstatementlist" delimiter=";" keepdelimiter="NO"/>
+            <file name="mssql\\${parentArtifactId}\001\drop_constraint.sql" type="sqlstatementlist" delimiter=";" keepdelimiter="NO"/>
         </drop_constraint>
         <drop_index>
-            <file name="mssql\\${parentArtifactId}\002\drop_index.sql" type="sqlstatementlist" delimiter=";" keepdelimiter="NO"/>
+            <file name="mssql\\${parentArtifactId}\001\drop_index.sql" type="sqlstatementlist" delimiter=";" keepdelimiter="NO"/>
         </drop_index>
     </current>
