+++ jar-component-archetype/src/main/resources/archetype-resources/__rootArtifactId__-config/src/main/config/properties/com/silverpeas/components/__rootArtifactId__/multilang/__ClassNamePrefix__Bundle_fr.properties
