+++ jar-component-archetype/src/main/resources/archetype-resources/__rootArtifactId__-config/src/main/config/properties/com/silverpeas/components/__rootArtifactId__/settings/__ClassNamePrefix__Icons.properties
