+++ jar-component-archetype/src/main/resources/archetype-resources/__rootArtifactId__-config/src/main/config/xmlcomponents/__rootArtifactId__.xml
@@ -47,7 +47,7 @@
   </suite>
   <visible>true</visible>
   <portlet>false</portlet>
-  <instanceClassName>${package}.${className}Instanciator</instanceClassName>
+  <instanceClassName>${package}.${ClassNamePrefix}Instanciator</instanceClassName>
 <profiles>
     <profile name="admin">
       <label>
