+++ jar-component-archetype/src/main/resources/archetype-resources/__rootArtifactId__-war/src/main/webapp/__rootArtifactId__/jsp/welcome.jsp
@@ -30,25 +30,28 @@
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
 <%@ taglib uri="http://java.sun.com/jsp/jstl/fmt" prefix="fmt"%>
 <%@ taglib uri="http://www.silverpeas.com/tld/viewGenerator" prefix="view"%>
+
+<fmt:setLocale value="${symbol_dollar}{sessionScope[sessionController].language}" />
+<view:setBundle bundle="${symbol_dollar}{requestScope.resources.multilangBundle}" />
+<view:setBundle bundle="${symbol_dollar}{requestScope.resources.iconsBundle}" var="icons" />
+<c:set var="componentId" value="${symbol_dollar}{requestScope.componentId}" />
+<c:set var="sessionController">Silverpeas_${ClassNamePrefix}_<c:out value="${symbol_dollar}{componentId}" /></c:set>
+<c:set var="browseContext" value="${symbol_dollar}{requestScope.browseContext}" />
+
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Frameset//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-frameset.dtd">
 <html>
   <head>
     <view:looknfeel />
   </head>
-  <c:set var="componentId" value="${symbol_dollar}{requestScope.componentId}" />
-  <c:set var="sessionController">Silverpeas_${className}_<c:out value="${symbol_dollar}{componentId}" /></c:set>
-  <fmt:setLocale value="${symbol_dollar}{sessionScope[sessionController].language}" />
-  <view:setBundle bundle="${symbol_dollar}{requestScope.resources.multilangBundle}" />
-  <view:setBundle bundle="${symbol_dollar}{requestScope.resources.iconsBundle}" var="icons" />
-  <c:set var="browseContext" value="${symbol_dollar}{requestScope.browseContext}" />
-  <body bgcolor="#ffffff" leftmargin="5" topmargin="5" marginwidth="5" marginheight="5">
+  <body>
     <view:window>
       <view:frame>
-        Bienvenue sur le composant ${rootArtifactId}.
+        Bienvenue sur l'application ${rootArtifactId}.
       </view:frame>
       <view:frame>
         <view:board>
-          Cette instance s'appele <b><c:out value="${symbol_dollar}{browseContext[1]}" /></b>.<br/>
-          Elle se situe dans l'espace <b><c:out value="${symbol_dollar}{browseContext[0]}" /></b>.
+          Cette instance s'appelle <span class="${rootArtifactId}Name"><c:out value="${symbol_dollar}{browseContext[1]}" /></span>.<br/>
+          Elle se situe dans l'espace <span class="${rootArtifactId}Name"><c:out value="${symbol_dollar}{browseContext[0]}" /></span>.
         </view:board>
       </view:frame>
     </view:window>
